requirements.txt
asgiref==3.8.1
cachetools==5.5.1
Django==5.1.6
django-rql==4.4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
lark-parser==0.11.0
lib-rql==2.0.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1